backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "django>=5.2,<6.0",
    "django-cors-headers>=4.4.0",
    "django-filter>=24.0",
    "django-mptt>=0.18.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "flake8>=7.0.0",
    "pillow>=10.0.0",
    "psycopg[binary]>=3.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
